Fix RSpec/LeakyConstantDeclaration
