+++ spec/delayed/job_groups/plugin_spec.rb
@@ -5,56 +5,56 @@ describe Delayed::JobGroups::Plugin do
     @old_max_attempts = Delayed::Worker.max_attempts
     Delayed::Worker.max_attempts = 2
 
-    CompletionJob.invoked = false
-    CancellationJob.invoked = false
+    TestJobs::CompletionJob.invoked = false
+    TestJobs::CancellationJob.invoked = false
   end
 
   after do
     Delayed::Worker.max_attempts = @old_max_attempts
   end
 
-  let!(:job_group) { Delayed::JobGroups::JobGroup.create!(on_completion_job: CompletionJob.new) }
+  let!(:job_group) { Delayed::JobGroups::JobGroup.create!(on_completion_job: TestJobs::CompletionJob.new) }
 
   it "runs the completion job after completing other jobs" do
-    job_group.enqueue(NoOpJob.new)
-    job_group.enqueue(NoOpJob.new)
+    job_group.enqueue(TestJobs::NoOpJob.new)
+    job_group.enqueue(TestJobs::NoOpJob.new)
     job_group.mark_queueing_complete
     expect(job_group_count).to eq 1
     expect(queued_job_count).to eq 2
 
     # Run our first job
     Delayed::Worker.new.work_off(1)
-    expect(CompletionJob.invoked).to be(false)
+    expect(TestJobs::CompletionJob.invoked).to be(false)
     expect(job_group_count).to eq 1
     expect(queued_job_count).to eq 1
 
     # Run our second job which should enqueue the completion job
     Delayed::Worker.new.work_off(1)
-    expect(CompletionJob.invoked).to be(false)
+    expect(TestJobs::CompletionJob.invoked).to be(false)
     expect(job_group_count).to eq 0
     expect(queued_job_count).to eq 1
 
     # Now we should run the completion job
     Delayed::Worker.new.work_off(1)
-    expect(CompletionJob.invoked).to be(true)
+    expect(TestJobs::CompletionJob.invoked).to be(true)
     expect(queued_job_count).to eq 0
   end
 
   it "only runs the completion job after queueing is completed" do
-    job_group.enqueue(NoOpJob.new)
-    job_group.enqueue(NoOpJob.new)
+    job_group.enqueue(TestJobs::NoOpJob.new)
+    job_group.enqueue(TestJobs::NoOpJob.new)
     expect(job_group_count).to eq 1
     expect(queued_job_count).to eq 2
 
     # Run our first job
     Delayed::Worker.new.work_off(1)
-    expect(CompletionJob.invoked).to be(false)
+    expect(TestJobs::CompletionJob.invoked).to be(false)
     expect(job_group_count).to eq 1
     expect(queued_job_count).to eq 1
 
     # Run our second job
     Delayed::Worker.new.work_off(1)
-    expect(CompletionJob.invoked).to be(false)
+    expect(TestJobs::CompletionJob.invoked).to be(false)
     expect(job_group_count).to eq 1
     expect(queued_job_count).to eq 0
 
@@ -65,7 +65,7 @@ describe Delayed::JobGroups::Plugin do
 
     # Now we should run the completion job
     Delayed::Worker.new.work_off(1)
-    expect(CompletionJob.invoked).to be(true)
+    expect(TestJobs::CompletionJob.invoked).to be(true)
     expect(queued_job_count).to eq 0
   end
 
@@ -75,8 +75,8 @@ describe Delayed::JobGroups::Plugin do
       it "cancels the group" do
         Delayed::Worker.max_attempts = 1
 
-        job_group.enqueue(FailingJob.new)
-        job_group.enqueue(NoOpJob.new)
+        job_group.enqueue(TestJobs::FailingJob.new)
+        job_group.enqueue(TestJobs::NoOpJob.new)
         job_group.mark_queueing_complete
         expect(queued_job_count).to eq 2
         expect(job_group_count).to eq 1
@@ -84,7 +84,7 @@ describe Delayed::JobGroups::Plugin do
         # Run the job which should fail and cancel the JobGroup
         Delayed::Worker.new.work_off(1)
         # Completion job is not invoked
-        expect(CompletionJob.invoked).to be(false)
+        expect(TestJobs::CompletionJob.invoked).to be(false)
         expect(failed_job_count).to eq 1
         expect(queued_job_count).to eq 0
         expect(job_group_count).to eq 0
@@ -98,15 +98,15 @@ describe Delayed::JobGroups::Plugin do
       it "does not cancel the group" do
         Delayed::Worker.max_attempts = 1
 
-        job_group.enqueue(FailingJob.new)
-        job_group.enqueue(NoOpJob.new)
+        job_group.enqueue(TestJobs::FailingJob.new)
+        job_group.enqueue(TestJobs::NoOpJob.new)
         job_group.mark_queueing_complete
         expect(queued_job_count).to eq 2
         expect(job_group_count).to eq 1
 
         # Run the job which should fail don't cancel the JobGroup
         Delayed::Worker.new.work_off(1)
-        expect(CancellationJob.invoked).to be(false)
+        expect(TestJobs::CancellationJob.invoked).to be(false)
         expect(failed_job_count).to eq 1
         expect(queued_job_count).to eq 1
         expect(job_group_count).to eq 1
@@ -120,7 +120,7 @@ describe Delayed::JobGroups::Plugin do
         # Run the completion job
         Delayed::Worker.new.work_off(1)
         # Completion job is invoked
-        expect(CompletionJob.invoked).to be(true)
+        expect(TestJobs::CompletionJob.invoked).to be(true)
         expect(failed_job_count).to eq 1
         expect(queued_job_count).to eq 0
         expect(job_group_count).to eq 0
@@ -129,8 +129,8 @@ describe Delayed::JobGroups::Plugin do
       it "runs completion job if last job failed" do
         Delayed::Worker.max_attempts = 2
 
-        job_group.enqueue(NoOpJob.new)
-        job_group.enqueue(FailingJob.new)
+        job_group.enqueue(TestJobs::NoOpJob.new)
+        job_group.enqueue(TestJobs::FailingJob.new)
         job_group.mark_queueing_complete
         expect(queued_job_count).to eq 2
         expect(job_group_count).to eq 1
@@ -144,7 +144,7 @@ describe Delayed::JobGroups::Plugin do
         # Run the job which should error
         Delayed::Worker.new.work_off(1)
         # Completion job is not invoked
-        expect(CompletionJob.invoked).to be(false)
+        expect(TestJobs::CompletionJob.invoked).to be(false)
         expect(failed_job_count).to eq 0
         expect(queued_job_count).to eq 1
         expect(job_group_count).to eq 1
@@ -159,7 +159,7 @@ describe Delayed::JobGroups::Plugin do
         # Run the completion job
         Delayed::Worker.new.work_off(1)
         # Completion job is invoked
-        expect(CompletionJob.invoked).to be(true)
+        expect(TestJobs::CompletionJob.invoked).to be(true)
         expect(failed_job_count).to eq 1
         expect(queued_job_count).to eq 0
         expect(job_group_count).to eq 0
@@ -169,7 +169,7 @@ describe Delayed::JobGroups::Plugin do
 
   it "doesn't retry failed jobs if the job group has been canceled" do
     job_group.cancel
-    Delayed::Job.enqueue(FailingJob.new, job_group_id: job_group.id)
+    Delayed::Job.enqueue(TestJobs::FailingJob.new, job_group_id: job_group.id)
     expect(queued_job_count).to eq 1
 
     # Run the job which should fail and should not queue a retry
@@ -182,8 +182,8 @@ describe Delayed::JobGroups::Plugin do
     job_group.blocked = true
     job_group.save!
 
-    job_group.enqueue(NoOpJob.new)
-    job_group.enqueue(NoOpJob.new)
+    job_group.enqueue(TestJobs::NoOpJob.new)
+    job_group.enqueue(TestJobs::NoOpJob.new)
     job_group.mark_queueing_complete
     expect(Delayed::Job.count).to eq 2
 
@@ -197,41 +197,41 @@ describe Delayed::JobGroups::Plugin do
 
     # Run our first job
     Delayed::Worker.new.work_off(1)
-    expect(CompletionJob.invoked).to be(false)
+    expect(TestJobs::CompletionJob.invoked).to be(false)
     expect(job_group_count).to eq 1
     expect(Delayed::Job.count).to eq 1
 
     # Run our second job which should enqueue the completion job
     Delayed::Worker.new.work_off(1)
-    expect(CompletionJob.invoked).to be(false)
+    expect(TestJobs::CompletionJob.invoked).to be(false)
     expect(job_group_count).to eq 0
     expect(Delayed::Job.count).to eq 1
 
     # Now we should run the completion job
     Delayed::Worker.new.work_off(1)
-    expect(CompletionJob.invoked).to be(true)
+    expect(TestJobs::CompletionJob.invoked).to be(true)
     expect(Delayed::Job.count).to eq 0
   end
 
   context "when a cancellation job is provided" do
     let!(:job_group) do
-      Delayed::JobGroups::JobGroup.create!(on_completion_job: CompletionJob.new,
-                                           on_cancellation_job: CancellationJob.new)
+      Delayed::JobGroups::JobGroup.create!(on_completion_job: TestJobs::CompletionJob.new,
+                                           on_cancellation_job: TestJobs::CancellationJob.new)
     end
 
     it "runs the cancellation job after a job error causes cancellation" do
       Delayed::Worker.max_attempts = 1
 
-      job_group.enqueue(FailingJob.new)
-      job_group.enqueue(NoOpJob.new)
+      job_group.enqueue(TestJobs::FailingJob.new)
+      job_group.enqueue(TestJobs::NoOpJob.new)
       job_group.mark_queueing_complete
       expect(queued_job_count).to eq 2
       expect(job_group_count).to eq 1
 
       # Run the job which should fail and cancel the JobGroup
       Delayed::Worker.new.work_off(1)
-      expect(CompletionJob.invoked).to be(false)
-      expect(CancellationJob.invoked).to be(false)
+      expect(TestJobs::CompletionJob.invoked).to be(false)
+      expect(TestJobs::CancellationJob.invoked).to be(false)
       expect(failed_job_count).to eq 1
 
       expect(queued_job_count).to eq 1
@@ -239,24 +239,24 @@ describe Delayed::JobGroups::Plugin do
 
       # Now we should run the cancellation job
       Delayed::Worker.new.work_off(1)
-      expect(CompletionJob.invoked).to be(false)
-      expect(CancellationJob.invoked).to be(true)
+      expect(TestJobs::CompletionJob.invoked).to be(false)
+      expect(TestJobs::CancellationJob.invoked).to be(true)
       expect(queued_job_count).to eq 0
     end
 
     it "runs the cancellation job after the job group is cancelled" do
-      job_group.enqueue(NoOpJob.new)
-      job_group.enqueue(FailingJob.new)
+      job_group.enqueue(TestJobs::NoOpJob.new)
+      job_group.enqueue(TestJobs::FailingJob.new)
       job_group.mark_queueing_complete
       job_group.cancel
 
       # cancellation job should be queued
       expect(queued_job_count).to eq 1
-      expect(CancellationJob.invoked).to be(false)
+      expect(TestJobs::CancellationJob.invoked).to be(false)
 
       # Run the cancellation job
       Delayed::Worker.new.work_off(1)
-      expect(CancellationJob.invoked).to be(true)
+      expect(TestJobs::CancellationJob.invoked).to be(true)
       expect(queued_job_count).to eq 0
     end
   end
@@ -265,8 +265,8 @@ describe Delayed::JobGroups::Plugin do
     let!(:job_group) {  Delayed::JobGroups::JobGroup.create! }
 
     it "doesn't queue a non-existent completion job" do
-      job_group.enqueue(NoOpJob.new)
-      job_group.enqueue(NoOpJob.new)
+      job_group.enqueue(TestJobs::NoOpJob.new)
+      job_group.enqueue(TestJobs::NoOpJob.new)
       job_group.mark_queueing_complete
       expect(job_group_count).to eq 1
       expect(queued_job_count).to eq 2
@@ -283,37 +283,6 @@ describe Delayed::JobGroups::Plugin do
       expect(job_group_count).to eq 0
       expect(queued_job_count).to eq 0
       expect(failed_job_count).to eq 0
-    end
-  end
-
-  class FailingJob
-
-    def perform
-      raise 'Test failure'
-    end
-
-  end
-
-  class NoOpJob
-
-    def perform
-
-    end
-  end
-
-  class CompletionJob
-    cattr_accessor :invoked
-
-    def perform
-      CompletionJob.invoked = true
-    end
-  end
-
-  class CancellationJob
-    cattr_accessor :invoked
-
-    def perform
-      CancellationJob.invoked = true
     end
   end
